Fixes for SNAP-2707
- Added code to check if table exists or not in catalog.

+++ core/src/main/scala/org/apache/spark/sql/execution/ddl.scala
@@ -97,7 +97,8 @@ private[sql] case class DropTableOrViewCommand(isView: Boolean, ifExists: Boolea
     // check for table/view
     val qualifiedName = catalog.newQualifiedTableName(tableIdent)
     if (isView) {
-      if (!catalog.isView(qualifiedName) && !catalog.isLocalTemporaryView(qualifiedName)) {
+      if (!catalog.isView(qualifiedName) && !catalog.isLocalTemporaryView(qualifiedName)
+          && catalog.tableExists(qualifiedName)) {
         throw new AnalysisException(
           s"Cannot drop a table with DROP VIEW. Please use DROP TABLE instead")
       }
